tensormap/backend/requirements-core.txt
# API + DB + Socket.IO. Default DATABASE_URL is SQLite — no PostgreSQL driver needed.
# PostgreSQL: pip install -r requirements-postgres.txt and set DATABASE_URL=postgresql://...
#
# Use ONE Python environment. If you use `conda create ... python=3.12`, do not also
# activate a nested .venv built with Python 3.13+ — `pip` would use the wrong version.
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
sqlmodel==0.0.14
alembic==1.12.1
python-multipart==0.0.6
scikit-learn==1.3.2
# Training reports: PDF (import name `fpdf` from PyPI package `fpdf2`) + Word (`python-docx` → `import docx`)
fpdf2==2.8.2
python-docx==1.2.0
